cambiar precio a Long

+++ src/main/java/com/ventas/ventas/dto/ProductoDto.java
@@ -12,7 +12,7 @@ public class ProductoDto {
     
     private String nombre;
 
-    private Double precio;
+    private Long precio;
 
     private int stock;
